fix: 🐛 attribute NaN errors when loading encoded data
