+++ packages/upset/src/atoms/attributeAtom.ts
@@ -13,8 +13,12 @@ export const attributeValuesSelector = selectorFamily<number[], string>({
     (attribute: string) => ({ get }) => {
       const items = get(itemsAtom);
       const values = Object.values(items).map(
-        (val) => val[attribute] as number,
+        (val) => {
+          const retVal = val[attribute] as number;
+          return Number.isNaN(retVal) ? 0 : retVal;
+        },
       );
+
       return values;
     },
 });
@@ -27,6 +31,13 @@ export const attributeMinMaxSelector = selectorFamily<
   get:
     (attribute: string) => ({ get }) => {
       const values = get(attributeValuesSelector(attribute));
+      if (values.includes(NaN)) {
+        return {
+          min: 0,
+          max: 0,
+        };
+      }
+
       return {
         min: Math.min(...values),
         max: Math.max(...values),
